# mojefinance-service/api-model/src/main/resources/swagger/mojefinance-api.yaml
openapi: "3.0.0"
info:
  title: MojeFinance application
  version: 0.0.1

paths:
  /healthcheck:
    get:
      summary: Healthcheck
      operationId: healthcheck
      responses:
        '200':
          description: OK

  /banks:
    get:
      summary: Get connected banks
      operationId: getConnectedBanks
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectedBanksResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /banks/connect/ceska-sporitelna:
    get:
      summary: Connect to Ceska Sporitelna bank
      operationId: connectCeskaSporitelna
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code received from Ceska Sporitelna after user authorization
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /banks/connect/csob:
    get:
      summary: Connect to CSOB
      operationId: connectCSOB
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code received from CSOB after user authorization
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /banks/connect/kb:
    get:
      summary: Connect to KB
      operationId: connectKB
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code received from KB after user authorization
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /banks/connect/air-bank:
    get:
      summary: Connect to Air Bank
      operationId: connectAirBank
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code received from Air Bank after user authorization
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /banks/connect/reiffeisen-bank:
    get:
      summary: Connect to Reiffeisen Bank
      operationId: connectReiffeisenBank
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /registration:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPrivate'
      responses:
        '201':
          description: CREATED
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /profile:
    get:
      summary: Retrieve current user information
      operationId: getCurrentUser
      parameters:
        - $ref: '#/components/parameters/Authorization'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    Authorization:
      in: header
      name: Authorization
      required: true
      schema:
        type: string

  schemas:
    ConnectedBanksResponse:
      type: object
      properties:
        connectedBanks:
          type: array
          items:
            $ref: '#/components/schemas/Bank'

    Bank:
      type: object
      properties:
          bankName:
            type: string
          logoUrl:
            type: string
          bankConnectionStatus:
            type: string

    ProfileResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserPrivate'

    UserPublic:
      description: Public information about user
      type: object
      properties:
        username:
          type: string

    UserPrivate:
      description: Private information about user
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string

    ErrorResponse:
      type: object
      properties:
        message:
          type: string